Move to Visual Studio 2017 / CUDA Toolkit 9.2
-Changed compiler/IDE to Visual Studio 2017
-Changed CUDA toolkit to 9.2
-From now on, Windows builds will be **64-bit only**

+++ KeyFinderLib/KeyFinder.h
@@ -37,7 +37,7 @@ typedef struct {
 	std::string deviceName;
 	unsigned long long freeMemory;
 	unsigned long long deviceMemory;
-	unsigned int targets;
+	size_t targets;
 }KeyFinderStatusInfo;
